+++ src/main/java/Solution1.java
