+++ src/main/java/Solution2.java
