+++ src/main/java/Solution3.java
