+++ src/main/java/Solution6.java
