+++ src/main/java/Solution7.java
